pikasso_test/requirements.txt
django==4.2.5
djangorestframework==3.14.0
celery==5.3.4 